test with copper market data

+++ dzvector.cpp
@@ -3,17 +3,17 @@
 dzvector::dzvector()
 {
 	int i;
-	x = (double *) malloc(sizeof(double) * NO_DATA);
+	x = (long double *) malloc(sizeof(long double) * NO_DATA);
 	for (i = 0; i < NO_DATA; i++) 
 	{
 		x[i] = 0;
 	}
 }
 
-dzvector::dzvector(double *y)
+dzvector::dzvector(long double *y)
 {
 	int i;
-	x = (double *) malloc(sizeof(double) * NO_DATA);
+	x = (long double *) malloc(sizeof(long double) * NO_DATA);
 	for (i = 0; i < NO_DATA; i++)
 	{
 		x[i] = y[i];
@@ -30,14 +30,14 @@ void dzvector::output_vec()
 	cout << x[NO_DATA - 1] << endl;
 }
 
-double dzvector::get(int i)
+long double dzvector::get(int i)
 {
 	return x[i];
 }
 
-double dzvector::dot_product(dzvector* vec)
+long double dzvector::dot_product(dzvector* vec)
 {
-	double r;
+	long double r;
 	int i;
 	r = 0;
 	for(i = 0; i < NO_DATA; i++)
